feat: react-router-domをinstall 作成済みボード表示部分をコンポーネント化
