+++ package.json
@@ -53,6 +53,7 @@
     "react-app-polyfill": "^1.0.6",
     "react-dev-utils": "^10.2.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.1.2",
     "redux": "^4.0.5",
     "resolve": "1.15.0",
     "resolve-url-loader": "3.1.1",
@@ -144,6 +145,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-router-dom": "^5.1.3",
     "node-sass": "^4.13.1"
   }
 }
